Move things in conftest
