+++ tests/system_tests/test_blueapi_device_system.py
@@ -11,49 +11,51 @@ HOWTO:
        (or pytest tests/system_tests)
 """
 
-import time
+# import time
 from pathlib import Path
 
 import pytest
 from blueapi.client.client import BlueapiClient
-from blueapi.config import ApplicationConfig, RestConfig
-from blueapi.service.model import DeviceResponse, PlanResponse
-from pydantic import HttpUrl, TypeAdapter
-from requests.exceptions import ConnectionError
 
+# from blueapi.config import ApplicationConfig, RestConfig
+from blueapi.service.model import DeviceResponse, PlanResponse
+from pydantic import TypeAdapter
+
+# from pydantic import HttpUrl, TypeAdapter
+# from requests.exceptions import ConnectionError
 from .blueapi_system.example_devices import (
     AccessControlledOpticsMotors,
     FakeOpticsMotors,
     MotorPosition,
 )
 
-CONFIG = ApplicationConfig(api=RestConfig(url=HttpUrl("http://localhost:12345")))
+# CONFIG = ApplicationConfig(api=RestConfig(url=HttpUrl("http://localhost:12345")))
 
 
-@pytest.fixture
-def blueapi_client() -> BlueapiClient:
-    return BlueapiClient.from_config(config=CONFIG)
+# @pytest.fixture
+# def blueapi_client() -> BlueapiClient:
+#     return BlueapiClient.from_config(config=CONFIG)
 
 
-@pytest.fixture(scope="module", autouse=True)
-def wait_for_server():
-    client = BlueapiClient.from_config(config=CONFIG)
+# @pytest.fixture(scope="module", autouse=True)
+# def wait_for_server():
+#     client = BlueapiClient.from_config(config=CONFIG)
 
-    for _ in range(20):
-        try:
-            client.get_environment()
-            return
-        except ConnectionError:
-            ...
-        time.sleep(0.5)
-    raise TimeoutError("No connection to blueapi server")
+#     for _ in range(20):
+#         try:
+#             client.get_environment()
+#             return
+#         except ConnectionError:
+#             ...
+#         time.sleep(0.5)
+#     raise TimeoutError("No connection to blueapi server")
 
 
-@pytest.fixture(autouse=True)
-def clean_existing_tasks(blueapi_client: BlueapiClient):
-    for task in blueapi_client.get_all_tasks().tasks:
-        blueapi_client.clear_task(task.task_id)
-    yield
+# @pytest.fixture(autouse=True)
+# def clean_existing_tasks(blueapi_client: BlueapiClient):
+#     for task in blueapi_client.get_all_tasks().tasks:
+#         blueapi_client.clear_task(task.task_id)
+#     yield
 
 
 @pytest.mark.system_test
